WadPaths folded into PathsManager
